# src/modules/apidoc/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Media API",
    "description": "This is a simple API for social media and this is a test api document this project is not api base project this is a test project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4002"
    }
  ],
       "components": {
          "securitySchemes":{
               "bearerAuth": {
                  "type": "http"
               }
          }  
       },
...
  "paths": {
...
"/pages/{pageID}": {
  "get": {
       "tags": ["Post"],
       "summary": "Get all posts of a page",
       "parameters": [
         {
              "name": "pageID",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
         }
       ],
       "security": [
         {
              "bearerAuth": []
         }
       ],
       "responses": {
         "200": {
              "description": "Posts fetched successfully"
         },
         "500": {
              "description": "Internal server error"
         }
       }
  }
},
...
}
}